Split testData into smaller files ### Context testData.ts has become big and hard to maintain
### Fixes
Split testData into data, indexes and collections for better maintainability.

[ghstack-poisoned]
